added ownership of the listing, manage listing page

+++ app/Http/Controllers/ListingController.php
@@ -5,36 +5,30 @@ namespace App\Http\Controllers;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
-use Illuminate\Routing\Controller;
 
 class ListingController extends Controller
 {
-    public function index()
-    {
+    // Show all listings
+    public function index() {
         return view('listings.index', [
             'listings' => Listing::latest()->filter(request(['tag', 'search']))->paginate(6)
         ]);
     }
 
-    // Show single listing
-    public function show(Listing $listing)
-    {
+    //Show single listing
+    public function show(Listing $listing) {
         return view('listings.show', [
             'listing' => $listing
         ]);
     }
 
     // Show Create Form
-    public function create()
-    {
+    public function create() {
         return view('listings.create');
     }
 
-    // Store Form
-    public function store(Request $request)
-    {
-        // dd($request->file('logo'));
-        // dd($request);
+    // Store Listing Data
+    public function store(Request $request) {
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required', Rule::unique('listings', 'company')],
@@ -42,11 +36,10 @@ class ListingController extends Controller
             'website' => 'required',
             'email' => ['required', 'email'],
             'tags' => 'required',
-            'description' => 'required',
+            'description' => 'required'
         ]);
-        // dd($request);
 
-        if($request->hasFile('logo')){
+        if($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
@@ -57,15 +50,18 @@ class ListingController extends Controller
         return redirect('/')->with('message', 'Listing created successfully!');
     }
 
+    // Show Edit Form
     public function edit(Listing $listing) {
         return view('listings.edit', ['listing' => $listing]);
     }
 
     // Update Listing Data
-    public function update(Request $request, Listing $listing)
-    {
-        // dd($request->file('logo'));
-        // dd($request);
+    public function update(Request $request, Listing $listing) {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -73,26 +69,31 @@ class ListingController extends Controller
             'website' => 'required',
             'email' => ['required', 'email'],
             'tags' => 'required',
-            'description' => 'required',
-            'user_id' => 'required'
+            'description' => 'required'
         ]);
-        // dd($request);
 
-        if($request->hasFile('logo')){
+        if($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
-        $formFields['user_id'] = auth()->id();
-
         $listing->update($formFields);
 
-        return redirect()->back()->with('message', 'Listing updated successfully!');
+        return back()->with('message', 'Listing updated successfully!');
     }
 
     // Delete Listing
     public function destroy(Listing $listing) {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
     }
 
+    // Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listing()->get()]);
+    }
 }
